Added option to create header section
Give the possibility to the page writer to force creation of a header in a table. By specifying `header-rows="1"` in the datatable opening tag, this javascript will move the table body first row (or as many rows as specified) into the table head. Then DataTables can correctly transform the table.

This added functionality enables transforming tables which do not have a proper thead at creation, for example when created by CSV plugin.

Also slightly simplified the retrieval of `fixedHeaderOffsetTop` value.