Update script.js
Added two parameters colSearchType (which can be 'text' or 'list') and colSearchLoc (which can be 'top' or 'bot' or unspecified to default to bottom). This will allow for column level filtering via text/RegEx or a list of values in the column and allow for the column level filtering boxes to be at the top of bottom of the table.